+++ Assets/Plugins/iOS/Amplitude/AMPConstants.m
@@ -5,7 +5,7 @@
 
 NSString *const kAMPLibrary = @"amplitude-ios";
 NSString *const kAMPPlatform = @"iOS";
-NSString *const kAMPVersion = @"3.6.0";
+NSString *const kAMPVersion = @"3.8.3";
 NSString *const kAMPEventLogDomain = @"api.amplitude.com";
 NSString *const kAMPEventLogUrl = @"https://api.amplitude.com/";
 NSString *const kAMPDefaultInstance = @"$default_instance";
@@ -28,3 +28,9 @@ NSString *const AMP_OP_PREPEND = @"$prepend";
 NSString *const AMP_OP_SET = @"$set";
 NSString *const AMP_OP_SET_ONCE = @"$setOnce";
 NSString *const AMP_OP_UNSET = @"$unset";
+
+NSString *const AMP_REVENUE_PRODUCT_ID = @"$productId";
+NSString *const AMP_REVENUE_QUANTITY = @"$quantity";
+NSString *const AMP_REVENUE_PRICE = @"$price";
+NSString *const AMP_REVENUE_REVENUE_TYPE = @"$revenueType";
+NSString *const AMP_REVENUE_RECEIPT = @"$receipt";